+++ compiler-cli/src/dependencies.rs
@@ -20,6 +20,7 @@ use gleam_core::{
 };
 use hexpm::version::Version;
 use itertools::Itertools;
+use same_file::is_same_file;
 use smol_str::SmolStr;
 use strum::IntoEnumIterator;
 
@@ -511,7 +512,7 @@ fn get_manifest<Telem: Telemetry>(
     };
 
     if should_resolve {
-        let manifest = resolve_versions(runtime, mode, config, None, telemetry)?;
+        let manifest = resolve_versions(runtime, mode, paths, config, None, telemetry)?;
         return Ok((true, manifest));
     }
 
@@ -524,7 +525,7 @@ fn get_manifest<Telem: Telemetry>(
         Ok((false, manifest))
     } else {
         tracing::debug!("manifest_outdated");
-        let manifest = resolve_versions(runtime, mode, config, Some(&manifest), telemetry)?;
+        let manifest = resolve_versions(runtime, mode, paths, config, Some(&manifest), telemetry)?;
         Ok((true, manifest))
     }
 }
@@ -536,7 +537,7 @@ struct ProvidedPackage {
     requirements: HashMap<SmolStr, hexpm::version::Range>,
 }
 
-#[derive(Clone, Eq, PartialEq, Debug)]
+#[derive(Clone, Eq, Debug)]
 enum ProvidedPackageSource {
     Git { repo: SmolStr, commit: SmolStr },
     Local { path: PathBuf },
@@ -590,23 +591,46 @@ impl ProvidedPackage {
 impl ProvidedPackageSource {
     fn to_manifest_package_source(&self) -> ManifestPackageSource {
         match self {
-            ProvidedPackageSource::Git { repo, commit } => ManifestPackageSource::Git {
+            Self::Git { repo, commit } => ManifestPackageSource::Git {
                 repo: repo.clone(),
                 commit: commit.clone(),
             },
-            ProvidedPackageSource::Local { path } => {
-                ManifestPackageSource::Local { path: path.clone() }
-            }
+            Self::Local { path } => ManifestPackageSource::Local { path: path.clone() },
         }
     }
 
     fn to_toml(&self) -> String {
         match self {
-            ProvidedPackageSource::Git { repo, commit } => {
+            Self::Git { repo, commit } => {
                 format!(r#"{{ repo: "{}", commit: "{}" }}"#, repo, commit)
             }
-            ProvidedPackageSource::Local { path } => {
+            Self::Local { path } => {
                 format!(r#"{{ path: "{}" }}"#, path.to_string_lossy())
+            }
+        }
+    }
+}
+
+impl PartialEq for ProvidedPackageSource {
+    fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            (Self::Local { path: own_path }, Self::Local { path: other_path }) => {
+                is_same_file(own_path, other_path).unwrap_or(false)
+            }
+
+            (
+                Self::Git {
+                    repo: own_repo,
+                    commit: own_commit,
+                },
+                Self::Git {
+                    repo: other_repo,
+                    commit: other_commit,
+                },
+            ) => own_repo == other_repo && own_commit == other_commit,
+
+            (Self::Git { .. }, Self::Local { .. }) | (Self::Local { .. }, Self::Git { .. }) => {
+                false
             }
         }
     }
@@ -615,6 +639,7 @@ impl ProvidedPackageSource {
 fn resolve_versions<Telem: Telemetry>(
     runtime: tokio::runtime::Handle,
     mode: Mode,
+    project_paths: &ProjectPaths,
     config: &PackageConfig,
     manifest: Option<&Manifest>,
     telemetry: &Telem,
@@ -635,12 +660,16 @@ fn resolve_versions<Telem: Telemetry>(
             Requirement::Path { path } => provide_local_package(
                 SmolStr::from(&name),
                 &path,
+                project_paths,
                 &mut provided_packages,
                 &mut vec![],
             )?,
-            Requirement::Git { git } => {
-                provide_git_package(SmolStr::from(&name), &git, &mut provided_packages)?
-            }
+            Requirement::Git { git } => provide_git_package(
+                SmolStr::from(&name),
+                &git,
+                project_paths,
+                &mut provided_packages,
+            )?,
         };
         let _ = root_requirements.insert(name, version);
     }
@@ -678,19 +707,18 @@ fn resolve_versions<Telem: Telemetry>(
 fn provide_local_package(
     package_name: SmolStr,
     package_path: &Path,
+    project_paths: &ProjectPaths,
     provided: &mut HashMap<SmolStr, ProvidedPackage>,
     parents: &mut Vec<SmolStr>,
 ) -> Result<hexpm::version::Range> {
-    let canonical_path = package_path
-        .canonicalize()
-        .map_err(|_| Error::DependencyCanonicalizationFailed(package_name.to_string()))?;
     let package_source = ProvidedPackageSource::Local {
-        path: canonical_path.clone(),
+        path: package_path.to_path_buf(),
     };
     provide_package(
         package_name,
-        &canonical_path,
+        package_path,
         package_source,
+        project_paths,
         provided,
         parents,
     )
@@ -700,6 +728,7 @@ fn provide_local_package(
 fn provide_git_package(
     _package_name: SmolStr,
     _repo: &str,
+    _project_paths: &ProjectPaths,
     _provided: &mut HashMap<SmolStr, ProvidedPackage>,
 ) -> Result<hexpm::version::Range> {
     let _git = ProvidedPackageSource::Git {
@@ -714,6 +743,7 @@ fn provide_package(
     package_name: SmolStr,
     package_path: &Path,
     package_source: ProvidedPackageSource,
+    project_paths: &ProjectPaths,
     provided: &mut HashMap<SmolStr, ProvidedPackage>,
     parents: &mut Vec<SmolStr>,
 ) -> Result<hexpm::version::Range> {
@@ -770,9 +800,17 @@ fn provide_package(
                     package_path.join(path)
                 };
                 // Recursively walk local packages
-                provide_local_package(name.clone(), &resolved_path, provided, parents)?
+                provide_local_package(
+                    name.clone(),
+                    &resolved_path,
+                    project_paths,
+                    provided,
+                    parents,
+                )?
             }
-            Requirement::Git { git } => provide_git_package(name.clone(), &git, provided)?,
+            Requirement::Git { git } => {
+                provide_git_package(name.clone(), &git, project_paths, provided)?
+            }
         };
         let _ = requirements.insert(name, version);
     }
@@ -794,9 +832,11 @@ fn provide_package(
 #[test]
 fn provide_wrong_package() {
     let mut provided = HashMap::new();
+    let project_paths = crate::project_paths_at_current_directory();
     let result = provide_local_package(
         "wrong_name".into(),
-        &Path::new("./test/hello_world"),
+        Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -814,10 +854,12 @@ fn provide_wrong_package() {
 #[test]
 fn provide_existing_package() {
     let mut provided = HashMap::new();
+    let project_paths = crate::project_paths_at_current_directory();
 
     let result = provide_local_package(
         "hello_world".into(),
-        &Path::new("./test/hello_world"),
+        Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -828,7 +870,8 @@ fn provide_existing_package() {
 
     let result = provide_local_package(
         "hello_world".into(),
-        &Path::new("./test/hello_world"),
+        Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -841,10 +884,11 @@ fn provide_existing_package() {
 #[test]
 fn provide_conflicting_package() {
     let mut provided = HashMap::new();
-
+    let project_paths = crate::project_paths_at_current_directory();
     let result = provide_local_package(
         "hello_world".into(),
-        &Path::new("./test/hello_world"),
+        Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -855,10 +899,11 @@ fn provide_conflicting_package() {
 
     let result = provide_package(
         "hello_world".into(),
-        &Path::new("./test/other"),
+        Path::new("./test/other"),
         ProvidedPackageSource::Local {
             path: Path::new("./test/other").to_path_buf(),
         },
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -872,9 +917,11 @@ fn provide_conflicting_package() {
 #[test]
 fn provided_is_absolute() {
     let mut provided = HashMap::new();
+    let project_paths = crate::project_paths_at_current_directory();
     let result = provide_local_package(
         "hello_world".into(),
-        &Path::new("./test/hello_world"),
+        Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "subpackage".into()],
     );
@@ -884,7 +931,7 @@ fn provided_is_absolute() {
     );
     let package = provided.get("hello_world").unwrap().clone();
     if let ProvidedPackageSource::Local { path } = package.source {
-        assert!(path.is_absolute())
+        assert!(!path.is_absolute())
     } else {
         panic!("Provide_local_package provided a package that is not local!")
     }
@@ -893,9 +940,11 @@ fn provided_is_absolute() {
 #[test]
 fn provided_recursive() {
     let mut provided = HashMap::new();
+    let project_paths = crate::project_paths_at_current_directory();
     let result = provide_local_package(
         "hello_world".into(),
-        &Path::new("./test/hello_world"),
+        Path::new("./test/hello_world"),
+        &project_paths,
         &mut provided,
         &mut vec!["root".into(), "hello_world".into(), "subpackage".into()],
     );